Move handler related files to _handlers subdirectory
