+++ telegram/ext/__init__.py
@@ -59,36 +59,36 @@ __all__ = (
     "filters",
 )
 
-from . import filters
+from ._handlers import filters
+from ._handlers._basehandler import BaseHandler
+from ._handlers._callbackqueryhandler import CallbackQueryHandler
+from ._handlers._chatjoinrequesthandler import ChatJoinRequestHandler
+from ._handlers._chatmemberhandler import ChatMemberHandler
+from ._handlers._choseninlineresulthandler import ChosenInlineResultHandler
+from ._handlers._commandhandler import CommandHandler
+from ._handlers._pollanswerhandler import PollAnswerHandler
+from ._handlers._pollhandler import PollHandler
+from ._handlers._precheckoutqueryhandler import PreCheckoutQueryHandler
+from ._handlers._prefixhandler import PrefixHandler
+from ._handlers._shippingqueryhandler import ShippingQueryHandler
+from ._handlers._stringcommandhandler import StringCommandHandler
+from ._handlers._stringregexhandler import StringRegexHandler
+from ._handlers._typehandler import TypeHandler
+from ._handlers._inlinequeryhandler import InlineQueryHandler
+from ._handlers._conversationhandler import ConversationHandler
+from ._handlers._messagehandler import MessageHandler
 from ._aioratelimiter import AIORateLimiter
 from ._application import Application, ApplicationHandlerStop
 from ._applicationbuilder import ApplicationBuilder
-from ._basehandler import BaseHandler
 from ._basepersistence import BasePersistence, PersistenceInput
 from ._baseratelimiter import BaseRateLimiter
 from ._baseupdateprocessor import BaseUpdateProcessor, SimpleUpdateProcessor
 from ._callbackcontext import CallbackContext
 from ._callbackdatacache import CallbackDataCache, InvalidCallbackData
-from ._callbackqueryhandler import CallbackQueryHandler
-from ._chatjoinrequesthandler import ChatJoinRequestHandler
-from ._chatmemberhandler import ChatMemberHandler
-from ._choseninlineresulthandler import ChosenInlineResultHandler
-from ._commandhandler import CommandHandler
 from ._contexttypes import ContextTypes
-from ._conversationhandler import ConversationHandler
 from ._defaults import Defaults
 from ._dictpersistence import DictPersistence
 from ._extbot import ExtBot
-from ._inlinequeryhandler import InlineQueryHandler
 from ._jobqueue import Job, JobQueue
-from ._messagehandler import MessageHandler
 from ._picklepersistence import PicklePersistence
-from ._pollanswerhandler import PollAnswerHandler
-from ._pollhandler import PollHandler
-from ._precheckoutqueryhandler import PreCheckoutQueryHandler
-from ._prefixhandler import PrefixHandler
-from ._shippingqueryhandler import ShippingQueryHandler
-from ._stringcommandhandler import StringCommandHandler
-from ._stringregexhandler import StringRegexHandler
-from ._typehandler import TypeHandler
 from ._updater import Updater
