+++ telegram/ext/_handlers/_typehandler.py
@@ -16,35 +16,29 @@
 #
 # You should have received a copy of the GNU Lesser Public License
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the StringCommandHandler class."""
+"""This module contains the TypeHandler class."""
 
-from typing import TYPE_CHECKING, Any, List, Optional
+from typing import Optional, Type, TypeVar
 
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._basehandler import BaseHandler
-from telegram.ext._utils.types import CCT, RT, HandlerCallback
+from telegram.ext._handlers._basehandler import BaseHandler
+from telegram.ext._utils.types import CCT, HandlerCallback
 
-if TYPE_CHECKING:
-    from telegram.ext import Application
+RT = TypeVar("RT")
+UT = TypeVar("UT")
 
 
-class StringCommandHandler(BaseHandler[str, CCT]):
-    """Handler class to handle string commands. Commands are string updates that start with
-    ``/``. The handler will add a :obj:`list` to the
-    :class:`CallbackContext` named :attr:`CallbackContext.args`. It will contain a list of strings,
-    which is the text following the command split on single whitespace characters.
-
-    Note:
-        This handler is not used to handle Telegram :class:`telegram.Update`, but strings manually
-        put in the queue. For example to send messages with the bot using command line or API.
+class TypeHandler(BaseHandler[UT, CCT]):
+    """Handler class to handle updates of custom types.
 
     Warning:
         When setting :paramref:`block` to :obj:`False`, you cannot rely on adding custom
         attributes to :class:`telegram.ext.CallbackContext`. See its docs for more info.
 
     Args:
-        command (:obj:`str`): The command this handler should listen for.
+        type (:external:class:`type`): The :external:class:`type` of updates this handler should
+            process, as determined by :obj:`isinstance`
         callback (:term:`coroutine function`): The callback function for this handler. Will be
             called when :meth:`check_update` has determined that an update should be processed by
             this handler. Callback signature::
@@ -53,6 +47,8 @@ class StringCommandHandler(BaseHandler[str, CCT]):
 
             The return value of the callback is usually ignored except for the special case of
             :class:`telegram.ext.ConversationHandler`.
+        strict (:obj:`bool`, optional): Use ``type`` instead of :obj:`isinstance`.
+            Default is :obj:`False`.
         block (:obj:`bool`, optional): Determines whether the return value of the callback should
             be awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`. Defaults to :obj:`True`.
@@ -60,49 +56,40 @@ class StringCommandHandler(BaseHandler[str, CCT]):
             .. seealso:: :wiki:`Concurrency`
 
     Attributes:
-        command (:obj:`str`): The command this handler should listen for.
+        type (:external:class:`type`): The :external:class:`type` of updates this handler should
+            process.
         callback (:term:`coroutine function`): The callback function for this handler.
+        strict (:obj:`bool`): Use :external:class:`type` instead of :obj:`isinstance`. Default is
+            :obj:`False`.
         block (:obj:`bool`): Determines whether the return value of the callback should be
             awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`.
 
     """
 
-    __slots__ = ("command",)
+    __slots__ = ("type", "strict")
 
     def __init__(
         self,
-        command: str,
-        callback: HandlerCallback[str, CCT, RT],
+        type: Type[UT],  # pylint: disable=redefined-builtin
+        callback: HandlerCallback[UT, CCT, RT],
+        strict: bool = False,
         block: DVType[bool] = DEFAULT_TRUE,
     ):
         super().__init__(callback, block=block)
-        self.command: str = command
+        self.type: Type[UT] = type
+        self.strict: Optional[bool] = strict
 
-    def check_update(self, update: object) -> Optional[List[str]]:
+    def check_update(self, update: object) -> bool:
         """Determines whether an update should be passed to this handler's :attr:`callback`.
 
         Args:
-            update (:obj:`object`): The incoming update.
+            update (:obj:`object`): Incoming update.
 
         Returns:
-            List[:obj:`str`]: List containing the text command split on whitespace.
+            :obj:`bool`
 
         """
-        if isinstance(update, str) and update.startswith("/"):
-            args = update[1:].split(" ")
-            if args[0] == self.command:
-                return args[1:]
-        return None
-
-    def collect_additional_context(
-        self,
-        context: CCT,
-        update: str,  # skipcq: BAN-B301
-        application: "Application[Any, CCT, Any, Any, Any, Any]",  # skipcq: BAN-B301
-        check_result: Optional[List[str]],
-    ) -> None:
-        """Add text after the command to :attr:`CallbackContext.args` as list, split on single
-        whitespaces.
-        """
-        context.args = check_result
+        if not self.strict:
+            return isinstance(update, self.type)
+        return type(update) is self.type  # pylint: disable=unidiomatic-typecheck
